+++ pom.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>org.jetax</groupId>
+	<groupId>com.hmhco</groupId>
 	<artifactId>testcontainers-consul</artifactId>
-	<version>0.1.2-SNAPSHOT</version>
+	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
 	<scm>
-          <developerConnection>scm:git:git@github.com:hmhco/testcontainers-consul.git</developerConnection>
-          <tag>HEAD</tag>
+	    <developerConnection>scm:git:git@github.com:hmhco/testcontainers-consul.git</developerConnection>
+	    <tag>HEAD</tag>
   	</scm>
 	<properties>
-		<java.version>15</java.version>
-		<maven.compiler.source>15</maven.compiler.source>
-		<maven.compiler.target>15</maven.compiler.target>
+		<java.version>17</java.version>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
 	</properties>
 	<dependencies>
         <dependency>
@@ -33,18 +33,18 @@
 		<dependency>
 		    <groupId>org.testcontainers</groupId>
 		    <artifactId>testcontainers</artifactId>
-		    <version>1.15.2</version>
+		    <version>1.16.2</version>
 		</dependency>
 		<dependency>
 		    <groupId>org.testcontainers</groupId>
 		    <artifactId>junit-jupiter</artifactId>
-		    <version>1.15.2</version>
+		    <version>1.16.2</version>
 		    <scope>test</scope>
 		</dependency>
 		<dependency>
 		    <groupId>org.testcontainers</groupId>
 		    <artifactId>vault</artifactId>
-		    <version>1.15.2</version>
+		    <version>1.16.2</version>
 		    <scope>test</scope>
 		</dependency>		
 		<!-- end::tests[] -->
